feat: debugger
add a "Clear" button to the right of the "Trace Visualizer" button (but keep that new button aligned to the right). This would reset the state of the visualizer. Additionally, add an "X" button on the top right corner (right next to the timestamp) of each of the traces that would allow you to dismiss that particular trace).